requirements.txt
asgiref==3.2.10
backcall==0.2.0
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.2
defusedxml==0.6.0
Django==3.0
django-allauth==0.42.0
django-phonenumber-field==5.0.0
django-rest-auth==0.9.5
djangorestframework==3.11.1
drf-yasg==1.17.1
idna==2.10
inflection==0.5.1
ipython==7.18.1
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.17.2
Jinja2==2.11.2
MarkupSafe==1.1.1
oauthlib==3.1.0
packaging==20.4
parso==0.7.1
pexpect==4.8.0
phonenumbers==8.12.11
pickleshare==0.7.5
prompt-toolkit==3.0.7
ptyprocess==0.6.0
Pygments==2.7.1
pyparsing==2.4.7
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
traitlets==5.0.4
uritemplate==3.0.1
urllib3==1.25.10
wcwidth==0.2.5
